requirements.txt
asgiref==3.4.1
bleach==4.1.0
Django==3.2.8
django-material-admin==1.8.5
django-summernote==0.8.20.0
packaging==21.0
Pillow==8.4.0
pyparsing==3.0.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.4
webencodings==0.5.1
